requirements.txt
requests~=2.28.2
pytest
bs4~=0.0.1
lxml
numpy
Brotli
python-dotenv
setuptools==58.2.0
slimit
Js2Py~=0.74
ply==3.4
beautifulsoup4~=4.11.2